% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FTPRL}
\alias{FTPRL}
\title{Parameters of Follow The Proximal Regularized Leader Algorithms}
\description{
Parameters of Follow The Proximal Regularized Leader Algorithms
}
\details{
TODO
}
\section{Slots}{

 \describe{
   \item{\code{alpha}: Learning rate.}
   \item{\code{beta}: Regularization of initial learning rate.}
   \item{\code{lambda1}: $L_1$ regularization.}
   \item{\code{lambda2}: $L_2$ regularization.}
 }
}

