% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logloss}
\alias{logloss}
\title{Log Loss Evaluator}
\usage{
logloss(p, y, tol = 1e-06)
}
\arguments{
\item{p}{numeric vector. The predicted probability}

\item{y}{logical vector. The answer}

\item{tol}{numeric value. The value larger than $1 - tol$ will be shrinkage to $1 - tol$.
The value smaller than $tol$ will be shrinkage to $tol$.}
}
\description{
Log Loss Evaluator
}

